+++ src/ManifoldMakie.jl
@@ -6,11 +6,19 @@ using Manifolds: Sphere
 using Makie
 import Makie: scatter, scatter!
 
+# Generic Recipes
+include("generic/geodesics.jl")
+
+include("hyperbolic/hyperboloid.jl")
+include("hyperbolic/poincareball.jl")
+include("hyperbolic/poincarehalfspace.jl")
 include("sphere.jl")
-include("geodesics.jl")
 
 export Sphere
 export scatter, scatter!
 export geodesics, geodesics!
+export hyperboloidplot, hyperboloidplot!
+export poincareballbplot, poincareballbplot!
+export poincarehalfspaceplot, poincarehalfspaceplot!
 export scattergeodesics, scattergeodesics!
 end # module ManifoldMakie