+++ src/sphere.jl
@@ -2,11 +2,13 @@
     sphereplot(::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}; kwargs...)
 
 Draw the [`Sphere`](@extref `Manifolds.Sphere`)`(2)` as a (transparent) surface with an overlaid wireframe.
-Combine with [`scatter`](@extref Makie.scatter) to plot points or paths on it.
+This can be combined with
+* [`scatter`](@extref `Makie.scatter`)`(M, pts)` to plot points thereon
+* [`arrows3d`](@extref `Makie.arrows3d`)`(M, pts, vecs)` to plot tangent vectors
+* [`geodesics`](@ref)`(M, pst)` and [`scattergeodesics`](@ref)`(M, pst)` to draw geodesics
 
 ```julia
 fig, ax, p = sphereplot(Manifolds.Sphere(2))
-scatter!(ax, Manifolds.Sphere(2), pts)
 ```
 """
 @recipe SpherePlot (M,) begin
@@ -24,25 +26,23 @@ scatter!(ax, Manifolds.Sphere(2), pts)
     Makie.mixin_generic_plot_attributes()...
 end
 
-function Makie.plot!(plot::SpherePlot{<:Tuple{Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}}})
+function Makie.plot!(p::SpherePlot{<:Tuple{Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}}})
     # create a new compute edge
     # with [:M, :wires] as input nodes (these must already exist)
     # with :sphere_mesh as the output node (this will be created)
     # running the computation defined in the do ... end block
     # where :M is M and :wires is mapped to n
-    map!(plot.attributes, [:M, :wires], :sphere_mesh) do M, n
+    map!(p.attributes, [:M, :wires], :sphere_mesh) do M, n
         return GeometryBasics.uv_normal_mesh(Tessellation(Makie.Sphere(Point3f(0), 1.0f0), n))
     end
     # A solid surface
     mesh!(
-        plot, plot.sphere_mesh;
-        color = plot.surfacecolor,
-        alpha = plot.surfacealpha,
-        transparency = true,
+        p, p.sphere_mesh;
+        color = p.surfacecolor, alpha = p.surfacealpha, transparency = true,
     )
     # a wireframe atop
-    wireframe!(plot, plot.sphere_mesh; color = plot.wirecolor, linewidth = plot.wirewidth, transparency = true)
-    return plot
+    wireframe!(p, p.sphere_mesh; color = p.wirecolor, linewidth = p.wirewidth, transparency = true)
+    return p
 end
 
 #
@@ -50,26 +50,44 @@ end
 # Overwrite sphereplot (as a bit of a hack) to remove axes and use the nice default sphere from the docs?
 function sphereplot(
         M::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}};
-        size = (1024, 1024), backgroundcolor = :white, show_axis = false, kwargs...
+        size = (1024, 1024), backgroundcolor = :white, show_axis = false, aspect = :data, kwargs...
     )
-    fig = Figure(backgroundcolor = backgroundcolor, size = size, kwargs...)
-    ax = LScene(fig[1, 1], show_axis = show_axis, kwargs...)
+    fig = Figure(; backgroundcolor = backgroundcolor, size = size)
+    ax = Axis3(fig[1, 1], aspect = aspect)
+    if !show_axis
+        hidedecorations!(ax)
+        hidespines!(ax)
+    end
+    ax.azimuth = π / 4
     pl = sphereplot!(ax, M; kwargs...)
     return Makie.FigureAxisPlot(fig, ax, pl)
 end
 # For `scatter(M, pts)`, `lines(M, pts)`, `scatterlines(M, pts)`
 # (and any other PointBased plot) work on a manifold via this overload.
 # We do not have to transform the points
-function Makie.convert_arguments(P::Makie.PointBased, ::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}, pts)
+function Makie.convert_arguments(P::Makie.PointBased, ::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}, pts::V) where {V <: AbstractVector{<:AbstractVector}}
+    return Makie.convert_arguments(P, Point3f.(pts))
+end
+# we already have P3fs, just pass down
+function Makie.convert_arguments(P::Makie.PointBased, ::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}, pts::V) where {V <: AbstractVector{<:Point}}
     return Makie.convert_arguments(P, pts)
 end
 
 # For arrows3d(M, pts, vecs) we want to combine the classical scatter with arrows,
 # where we assume that vecs[i] is in the tangent space of pts[1]
-function Makie.convert_arguments(::Makie.ArrowLike, ::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}, pts, vecs)
+# (a) From Manifolds.jl
+function Makie.convert_arguments(
+        ::Makie.ArrowLike, ::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}, pts::V, vecs::W
+    ) where {V <: AbstractVector{<:AbstractVector}, W <: AbstractVector{<:AbstractVector}}
     #Not 100 % sure why the [1] is necessary, taken from conversions happening in arrows.jlr
     return (
-        convert_arguments(Makie.PointBased(), pts)[1],
-        convert_arguments(Makie.PointBased(), vecs)[1],
+        convert_arguments(Makie.PointBased(), Point3f.(pts))[1], convert_arguments(Makie.PointBased(), Vec3f.(vecs))[1],
+    )
+end
+# (b) already in Point3f
+function Makie.convert_arguments(::Makie.ArrowLike, ::Manifolds.Sphere{ℝ, Manifolds.TypeParameter{Tuple{2}}}, pts::V, vecs::W) where {V <: AbstractVector{<:Point}, W <: AbstractVector{<:Vec}}
+    #Not 100 % sure why the [1] is necessary, taken from conversions happening in arrows.jlr
+    return (
+        convert_arguments(Makie.PointBased(), pts)[1], convert_arguments(Makie.PointBased(), vecs)[1],
     )
 end