+++ Changelog.md
@@ -4,7 +4,7 @@ All notable Changes to the Julia package `ManifoldMakie.jl` are documented in th
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
+'1q
 ## [0.1.0] unreleased
 
 Initial release.
@@ -12,3 +12,6 @@ Initial release.
 ## Added
 
 * a `sphereplot(Manifolds.Sphere(2))` to start a plot with a sphere to put points on.
+* a `hyperboloidplot(Hyperbolic(n))` for `n=2,3` to start a plot in the hyperboloid model.
+* a `poincareballplot(Hyperbolic(n))` for `n=2,3` o start a plot in the Poincaré disc or ball model.
+* a `poincarehalfspaceplot(Hyperbolic(n))` for `n=2,3` to start a plot in the Poincaré hal plane or half space model.